+++ idea/testData/intentions/addNamesToCallArguments/allNamed.kt
@@ -1,0 +1,6 @@
+// IS_APPLICABLE: false
+fun foo(s: String, b: Boolean){}
+
+fun bar() {
+    <caret>foo(s = "", b = true)
+}