+++ idea/testData/intentions/addNamesToCallArguments/simple.kt
@@ -1,0 +1,5 @@
+fun foo(s: String, b: Boolean){}
+
+fun bar() {
+    <caret>foo("", true)
+}