+++ idea/testData/quickfix/increaseVisibility/privateMemberToInternalWithExposed.kt
@@ -1,5 +1,6 @@
 // "Make bar internal" "false"
 // ACTION: Convert property initializer to getter
+// ACTION: Add names to call arguments
 // ERROR: Cannot access 'bar': it is private in 'First'
 
 private data class Data(val x: Int)